Split `toc_vfs` into separate crates
