+++ compiler/toc_analysis/Cargo.toml
@@ -19,6 +19,7 @@ thiserror = "1.0.24"
 [dev-dependencies]
 toc_ast_db = { path = "../toc_ast_db" }
 toc_vfs = { path = "../toc_vfs" }
+toc_vfs_db = { path = "../toc_vfs_db" }
 
 insta = "1.7"
 unindent = "0.1.7"
