+++ compiler/toc_analysis/src/test_db.rs
@@ -7,9 +7,10 @@ use toc_ast_db::SourceGraph;
 use toc_hir::library::LibraryId;
 use toc_hir_db::db::HirDatabase;
 use toc_salsa::salsa;
+use toc_vfs_db::db::VfsDatabaseExt;
 
 #[salsa::database(
-    toc_vfs::db::FileSystemStorage,
+    toc_vfs_db::db::FileSystemStorage,
     toc_ast_db::db::SourceParserStorage,
     toc_hir_db::db::HirDatabaseStorage,
     crate::db::TypeInternStorage,
@@ -30,7 +31,8 @@ toc_vfs::impl_has_vfs!(TestDb, vfs);
 impl TestDb {
     pub(crate) fn from_source(source: &str) -> (Self, LibraryId) {
         let mut db = TestDb::default();
-        toc_vfs::generate_vfs(&mut db, source);
+        let fixture = toc_vfs::generate_vfs(&mut db, source);
+        db.insert_fixture(fixture);
 
         let root_file = db.vfs.intern_path("src/main.t".into());
         let mut source_graph = SourceGraph::default();
