+++ compiler/toc_driver/Cargo.toml
@@ -13,6 +13,7 @@ toc_hir = { path = "../toc_hir" }
 toc_hir_db = { path = "../toc_hir_db" }
 toc_hir_pretty = { path = "../toc_hir_pretty" }
 toc_vfs = { path = "../toc_vfs" }
+toc_vfs_db = { path = "../toc_vfs_db" }
 toc_analysis = { path = "../toc_analysis" }
 toc_span = { path = "../toc_span" }
 toc_ast_db = { path = "../toc_ast_db" }
