+++ compiler/toc_hir_lowering/Cargo.toml
@@ -18,6 +18,7 @@ indexmap = "1"
 [dev-dependencies]
 toc_salsa = { path = "../toc_salsa" }
 toc_vfs = { path = "../toc_vfs" }
+toc_vfs_db = { path = "../toc_vfs_db" }
 toc_hir_pretty = { path = "../toc_hir_pretty" }
 if_chain = "1.0"
 approx = "0.5"
