+++ compiler/toc_vfs/src/fixture.rs
@@ -1,13 +1,13 @@
 //! Test fixture generation
 
-use crate::{
-    db::{self, VfsDatabaseExt},
-    HasVfs, LoadError, LoadStatus,
-};
+use toc_span::FileId;
+
+use crate::{HasVfs, LoadResult};
+use crate::{LoadError, LoadStatus};
 
 pub const FILE_DELIMITER_START: &str = "%%-";
 
-/// Generates a VFS tree from the given text source
+/// Generates a VFS tree from the given text source, producing a set of [`FixtureFiles`]
 ///
 /// There are two kinds of text sources:
 ///
@@ -38,8 +38,12 @@ pub const FILE_DELIMITER_START: &str = "%%-";
 /// put "and i am in another"
 /// "#);
 /// ```
-pub fn generate_vfs<DB: db::FileSystem + HasVfs>(db: &mut DB, source: &str) {
-    // ???: How should we deal with inserting path resolutions
+#[must_use = "fixture files are not inserted into the given database (use `VfsDatabaseExt::insert_fixture`)"]
+pub fn generate_vfs<DB: HasVfs>(db: &mut DB, source: &str) -> FixtureFiles {
+    // ???: How we deal with inserting path resolutions
+    // We just do it here for direct paths, let source graph deal with the others
+
+    let mut fixture = FixtureFiles { files: vec![] };
 
     // Segment into lines
     let lines: Vec<_> = source.split('\n').collect();
@@ -79,9 +83,9 @@ pub fn generate_vfs<DB: db::FileSystem + HasVfs>(db: &mut DB, source: &str) {
                 // starts with the delimiter, proceed as normal
                 false
             } else if files.len() > 1 {
-                // is multifile, but with leading text present
+                // is multi-file, but with leading text present
                 // make sure that the leading text is just whitespace
-                panic!("text present before first file (`src/main.t` must be expicitly defined)");
+                panic!("text present before first file (`src/main.t` must be explicitly defined)");
             } else {
                 // is just a single file
                 true
@@ -91,13 +95,16 @@ pub fn generate_vfs<DB: db::FileSystem + HasVfs>(db: &mut DB, source: &str) {
             true
         }
     } else {
-        // Guaranateed to be a single file source, since it's empty
+        // Guaranteed to be a single file source, since it's empty
         true
     };
 
     if is_single_file {
-        db.insert_file("src/main.t", source);
-        return;
+        let main = db.get_vfs_mut().intern_path("src/main.t".into());
+        fixture
+            .files
+            .push((main, Ok(LoadStatus::Modified(source.into()))));
+        return fixture;
     }
 
     // Build up the files
@@ -143,59 +150,77 @@ pub fn generate_vfs<DB: db::FileSystem + HasVfs>(db: &mut DB, source: &str) {
             Ok(LoadStatus::Modified(source.into()))
         };
 
-        db.update_file(file, source);
+        fixture.files.push((file, source));
+    }
+
+    fixture
+}
+
+/// File tree generated by [`generate_vfs`]
+#[derive(Debug)]
+pub struct FixtureFiles {
+    pub files: Vec<(FileId, LoadResult)>,
+}
+
+impl FixtureFiles {
+    pub fn file_source(&self, file: FileId) -> (String, Option<LoadError>) {
+        self.files
+            .iter()
+            .find(|info| info.0 == file)
+            .map(|(_, res)| {
+                res.as_ref()
+                    .map(|sts| match sts {
+                        LoadStatus::Unchanged => unreachable!(),
+                        LoadStatus::Modified(src) => {
+                            (String::from_utf8_lossy(src).to_string(), None)
+                        }
+                    })
+                    .unwrap_or_else(|err| (String::new(), Some(err.clone())))
+            })
+            .unwrap()
     }
 }
 
 #[cfg(test)]
 mod test {
-    use toc_salsa::salsa;
-
-    use crate::{
-        db::{FileSystem, FileSystemStorage},
-        impl_has_vfs, Vfs,
-    };
+    use crate::{impl_has_vfs, Vfs};
 
     use super::*;
 
-    #[salsa::database(FileSystemStorage)]
     #[derive(Default)]
     struct VfsTestDB {
-        storage: salsa::Storage<Self>,
         vfs: Vfs,
     }
-
-    impl salsa::Database for VfsTestDB {}
 
     impl_has_vfs!(VfsTestDB, vfs);
 
     #[test]
     fn single_file() {
         let mut db = VfsTestDB::default();
+        let srcs = generate_vfs(&mut db, "single file, yay");
 
-        generate_vfs(&mut db, "single file, yay");
         let root_file = db.vfs.resolve_path(None, "src/main.t").into_file_id();
-        let source = db.file_source(root_file).0;
+        let source = srcs.file_source(root_file).0;
         assert_eq!(source.as_str(), "single file, yay");
     }
 
     #[test]
     fn single_file_empty() {
         let mut db = VfsTestDB::default();
+        let srcs = generate_vfs(&mut db, "");
 
-        generate_vfs(&mut db, "");
         let root_file = db.vfs.resolve_path(None, "src/main.t").into_file_id();
-        let res = db.file_source(root_file);
+        let res = srcs.file_source(root_file);
         assert_eq!((res.0.as_str(), res.1), ("", None));
     }
 
     #[test]
     fn single_file_many_lines() {
         let mut db = VfsTestDB::default();
+        let srcs = generate_vfs(&mut db, "single file\nbut multiple lines");
 
-        generate_vfs(&mut db, "single file\nbut multiple lines");
         let root_file = db.vfs.resolve_path(None, "src/main.t").into_file_id();
-        let source = db.file_source(root_file).0;
+        let source = srcs.file_source(root_file).0;
         assert_eq!(source.as_str(), "single file\nbut multiple lines");
     }
 
@@ -209,21 +234,22 @@ mod test {
         let mut db = VfsTestDB::default();
 
         // Leading whitespace is okay
-        generate_vfs(
+        let srcs = generate_vfs(
             &mut db,
             &format!(
                 "     \n    %%- src/main.t\n{0}\n%%- src/file1.t\n{1}\n%%- removed removed/ya.t\n%%- src/file2.t\n{2}",
                 FILE_SOURCES[0], FILE_SOURCES[1], FILE_SOURCES[2]
             ),
         );
+
         let file = db.vfs.resolve_path(None, "src/main.t").into_file_id();
-        assert_eq!(db.file_source(file).0.as_str(), FILE_SOURCES[0]);
+        assert_eq!(srcs.file_source(file).0.as_str(), FILE_SOURCES[0]);
         let file = db.vfs.resolve_path(None, "src/file1.t").into_file_id();
-        assert_eq!(db.file_source(file).0.as_str(), FILE_SOURCES[1]);
+        assert_eq!(srcs.file_source(file).0.as_str(), FILE_SOURCES[1]);
         let file = db.vfs.resolve_path(None, "src/file2.t").into_file_id();
-        assert_eq!(db.file_source(file).0.as_str(), FILE_SOURCES[2]);
+        assert_eq!(srcs.file_source(file).0.as_str(), FILE_SOURCES[2]);
         let file = db.vfs.resolve_path(None, "removed/ya.t").into_file_id();
-        let res = db.file_source(file);
+        let res = srcs.file_source(file);
         assert_eq!(
             (res.0.as_str(), res.1),
             ("", Some(crate::LoadError::NotFound))
@@ -234,20 +260,20 @@ mod test {
     fn multifile_one_file() {
         const FILE_SOURCES: &[&str] = &["singular_file"];
         let mut db = VfsTestDB::default();
+        let srcs = generate_vfs(&mut db, &format!("%%- one/file.t\n{0}", FILE_SOURCES[0]));
 
-        generate_vfs(&mut db, &format!("%%- one/file.t\n{0}", FILE_SOURCES[0]));
         let file = db.vfs.resolve_path(None, "one/file.t").into_file_id();
-        assert_eq!(db.file_source(file).0.as_str(), FILE_SOURCES[0]);
+        assert_eq!(srcs.file_source(file).0.as_str(), FILE_SOURCES[0]);
     }
 
     #[test]
     fn multifile_removed_as_path() {
         const FILE_SOURCES: &[&str] = &["not actually empty"];
         let mut db = VfsTestDB::default();
+        let srcs = generate_vfs(&mut db, &format!("%%- removed/ya.t\n{0}", FILE_SOURCES[0]));
 
-        generate_vfs(&mut db, &format!("%%- removed/ya.t\n{0}", FILE_SOURCES[0]));
         let file = db.vfs.resolve_path(None, "removed/ya.t").into_file_id();
-        let res = db.file_source(file);
+        let res = srcs.file_source(file);
         assert_eq!((res.0.as_str(), res.1), (FILE_SOURCES[0], None));
     }
 
@@ -255,22 +281,23 @@ mod test {
     fn multifile_empty_files() {
         let mut db = VfsTestDB::default();
 
-        generate_vfs(
+        let srcs = generate_vfs(
             &mut db,
             &format!("%%- empty/file0.t\n%%- empty/file1.t\n%%- empty/file2.t"),
         );
+
         let file = db.vfs.resolve_path(None, "empty/file0.t").into_file_id();
-        assert_eq!(db.file_source(file).0.as_str(), "");
+        assert_eq!(srcs.file_source(file).0.as_str(), "");
         let file = db.vfs.resolve_path(None, "empty/file1.t").into_file_id();
-        assert_eq!(db.file_source(file).0.as_str(), "");
+        assert_eq!(srcs.file_source(file).0.as_str(), "");
         let file = db.vfs.resolve_path(None, "empty/file2.t").into_file_id();
-        assert_eq!(db.file_source(file).0.as_str(), "");
+        assert_eq!(srcs.file_source(file).0.as_str(), "");
     }
 
     #[test]
-    #[should_panic = "text present before first file (`src/main.t` must be expicitly defined)"]
+    #[should_panic = "text present before first file (`src/main.t` must be explicitly defined)"]
     fn multifile_with_leading_text() {
         let mut db = VfsTestDB::default();
-        generate_vfs(&mut db, "        some text\n%%- src/a_file.t");
+        let _ = generate_vfs(&mut db, "        some text\n%%- src/a_file.t");
     }
 }
