+++ compiler/toc_vfs_db/Cargo.toml
@@ -1,12 +1,11 @@
 [package]
-name = "toc_vfs"
+name = "toc_vfs_db"
 version = "0.1.0"
-authors = ["DropDemBits <r3usrlnd@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-if_chain = "1.0.1"
-indexmap = "1.7.0"
-toc_span = { path =  "../toc_span" }
+toc_span = { path = "../toc_span" }
+toc_salsa = { path = "../toc_salsa" }
+toc_vfs = { path = "../toc_vfs" }
