+++ lsp-server/Cargo.toml
@@ -20,4 +20,5 @@ toc_ast_db = {path = "../compiler/toc_ast_db"}
 toc_reporting = {path = "../compiler/toc_reporting"}
 toc_span = {path = "../compiler/toc_span"}
 toc_vfs = {path = "../compiler/toc_vfs"}
+toc_vfs_db = {path = "../compiler/toc_vfs_db"}
 toc_salsa = { path = "../compiler/toc_salsa" }
