+++ lsp-server/src/main.rs
@@ -12,8 +12,8 @@ use toc_analysis::db::HirAnalysis;
 use toc_ast_db::db::{AstDatabaseExt, SourceParser, SpanMapping};
 use toc_ast_db::SourceGraph;
 use toc_salsa::salsa;
-use toc_vfs::db::VfsDatabaseExt;
 use toc_vfs::LoadStatus;
+use toc_vfs_db::db::VfsDatabaseExt;
 
 type DynError = Box<dyn Error + Sync + Send>;
 
@@ -220,7 +220,7 @@ impl IntoPosition for toc_ast_db::span::LspPosition {
 }
 
 #[salsa::database(
-    toc_vfs::db::FileSystemStorage,
+    toc_vfs_db::db::FileSystemStorage,
     toc_ast_db::db::SpanMappingStorage,
     toc_ast_db::db::SourceParserStorage,
     toc_hir_db::db::HirDatabaseStorage,
